requirements.txt
Twisted
txamqp
pyyaml
feedparser
xmpppy
jinja2
wokkel
pyopenssl
python-spidermonkey
python-mpd-twisted
http://www.freedesktop.org/~lanius/pyxdg-0.18.tar.gz
